% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cable_len}
\alias{cable_len}
\title{get cable length of neuron}
\usage{
cable_len(nl, skid)
}
\description{
get cable length of neuron
}
\details{
return cable length given a skeleton id
   and a neuronlist that contains the neuron,
}

