% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{check_tag}
\alias{check_tag}
\title{Check occurrences of a tag in each neuron of a neuron list}
\usage{
check_tag(nl, tag, counts = 1)
}
\arguments{
\item{nl}{a neuron list}

\item{tag}{name of the tag(e.g. 'pedunculus-calyx boundary')}

\item{counts}{number of occurrences of the tag in each neuron}
}
\value{
a list containing the number of tag occurrences in each neuron
}
\description{
Check occurrences of a tag in each neuron of a neuron list
}

