% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fetch_mb_neurons}
\alias{fetch_mb_neurons}
\title{download neurons associated with annotations}
\usage{
fetch_mb_neurons(annos, conn, anno_type = "annos", ref = "FCWB",
  border_tag = NULL, ...)
}
\arguments{
\item{annos}{a list of annotations used in CATMAID}

\item{border_tag}{border tag name (e.g 'bouton border' or 'claw border'),
if not (\code{NULL}) it will use the \code{\link{segment_arbor}} function to process the arbors}
}
\description{
transform the neurons if ref is supplied,
process their anatomy segments if border_tag is supplied
}
\details{
currently only works if each neuron is annotated with only one of the annotations
}
\seealso{
\code{\link{segment_arbor}}, \code{\link{fetchn_by_annotation}}
}

