% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fetchn_by_annotation}
\alias{fetchn_by_annotation}
\title{download neurons associated with an annotation}
\usage{
fetchn_by_annotation(anno, conn, anno_type, ...)
}
\arguments{
\item{anno}{an annotation used in CATMAID}
}
\description{
download neurons associated with an annotation
}
\details{
The annotation is attached in the meta-data in a field called anno_type.
Also include in each neuron a field called skids that contains skids. Skids serve
as a convenient id for the neuron.
}

