% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_tagged_node}
\alias{get_tagged_node}
\title{return attributes of nodes with a given tag}
\usage{
get_tagged_node(n, tag, rval)
}
\arguments{
\item{n}{a neuron object}

\item{tag}{name of the tag(e.g. 'pedunculus-calyx boundary')}

\item{return_fields}{type of output to return
("indices" or any combination of c("PointNo","Label","X","Y","Z","W","Parent"))}
}
\description{
return attributes of nodes with a given tag
}

