% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mushroom-package.R
\docType{package}
\name{mushroom-package}
\alias{mushroom}
\alias{mushroom-package}
\title{Analysis of Mushroom Body neurons using the R NeuroAnatomy Toolbox}
\description{
Analysis of Mushroom Body neurons using the R NeuroAnatomy Toolbox
}
\keyword{package}

