% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{segment_arbor}
\alias{segment_arbor}
\title{Find a neuron's anatomic segments delimited by border_tag}
\usage{
segment_arbor(n, border_tag, skid = NULL)
}
\arguments{
\item{n}{A neuron}

\item{border_tag}{name of the tag labeling the nodes that define the border}

\item{skid}{optional, but I always like to include skid as part of the unique
id for each anatomic segment.}
}
\value{
neuron with 2 new fields: AnaSeg, a list with each element containing
node indices in a single anatmic segment (a bouton or a claw); BoutonEnds, indices of end nodes of
each terminal bouton
}
\description{
Anatomic segments could be boutons ('bouton border' tag)
or claws ('claw border' tag).
The tags label border nodes that define the anatomic segments in
alternate fashion.
Take boutons for example in dfs order,
Root -> 1st tagged node: non-bouton,
1st tagged node -> 2nd tagged node: bouton,
2nd tagged node -> 3rd tagged node: non-bouton
...
}

