% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{skid_name}
\alias{skid_name}
\title{given a skid id and neuronlist return neuron name}
\usage{
skid_name(skid, nl)
}
\arguments{
\item{nl}{a neuron list}
}
\description{
given a skid id and neuronlist return neuron name
}

