% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{walk_skeleton}
\alias{walk_skeleton}
\title{a dfs search to traverse a neuron}
\usage{
walk_skeleton(x, starting_idx)
}
\arguments{
\item{starting_idx}{the node index of the starting point of the traversal}
}
\description{
a dfs search to traverse a neuron
}

