% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{z_to_node_idx}
\alias{z_to_node_idx}
\title{return an unique node index around a section number}
\usage{
z_to_node_idx(n, z)
}
\description{
return an unique node index around a section number
}
\details{
given a neuron and a section number, return a unique node
   index in the neuron closest to the section. Note it will look
   up and down until it finds an unique node (exclude no node or
   more than 1 node)
}

